+++ .idea/misc.xml
@@ -5,6 +5,7 @@
       <map>
         <entry key="..\:/Users/rajan/AndroidStudioProjects/inQuiRy/app/src/main/res/layout/activity_main.xml" value="0.335" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.41309255079006774" />
+        <entry key="app/src/main/res/layout/activity_scanner.xml" value="0.3658155289213132" />
       </map>
     </option>
   </component>
