+++ app/src/main/AndroidManifest.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.cmput301w22t13.inquiry">
-    <uses-permission android:name="android.permission.CAMERA"
-        />
 
+    <uses-permission android:name="android.permission.CAMERA" />
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
@@ -14,10 +13,13 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.InQuiRy">
         <activity
-            android:name=".Scanner"
+            android:name=".ScanActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".classes.Scanner"
+            android:exported="false" />
+        <activity
+            android:name=".classes.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
