--- app/src/main/java/com/cmput301w22t13/inquiry/LeaderBoard.java
@@ -1,5 +1,0 @@
-package com.cmput301w22t13.inquiry;
-
-public class LeaderBoard {
-
-}
