--- app/src/main/java/com/cmput301w22t13/inquiry/Owner.java
@@ -1,4 +1,0 @@
-package com.cmput301w22t13.inquiry;
-
-public class Owner {
-}
