+++ app/src/main/java/com/cmput301w22t13/inquiry/ScanActivity.java
@@ -1,0 +1,14 @@
+package com.cmput301w22t13.inquiry;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class ScanActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_scan);
+    }
+}
