+++ app/src/main/java/com/cmput301w22t13/inquiry/classes/LeaderBoard.java
@@ -1,0 +1,5 @@
+package com.cmput301w22t13.inquiry.classes;
+
+public class LeaderBoard {
+
+}
