+++ app/src/main/java/com/cmput301w22t13/inquiry/classes/Map.java
@@ -1,0 +1,4 @@
+package com.cmput301w22t13.inquiry.classes;
+
+public class Map {
+}
