+++ app/src/main/java/com/cmput301w22t13/inquiry/classes/Player.java
@@ -1,4 +1,4 @@
-package com.cmput301w22t13.inquiry;
+package com.cmput301w22t13.inquiry.classes;
 
 public class Player {
     private String userName;
