+++ app/src/main/java/com/cmput301w22t13/inquiry/classes/QRCode.java
@@ -1,4 +1,4 @@
-package com.cmput301w22t13.inquiry;
+package com.cmput301w22t13.inquiry.classes;
 
 public class QRCode {
     private String hash;
